made objects more consistent with framestrcuture
