+++ src/hdf5objects/arrayframes/__init__.py
@@ -1,5 +1,5 @@
 """ __init__.py
-The base dataclasses for this package.
+Implementations for HDF5 objects as frames from framestruct.
 """
 # Package Header #
 from ..header import *
@@ -10,7 +10,6 @@ __credits__ = __credits__
 __maintainer__ = __maintainer__
 __email__ = __email__
 
-
 # Imports #
 # Local Packages #
-from .dataobjects import *
+from .hdf5eegframe import HDF5EEGFrame
