--- src/hdf5objects/dataframes/__init__.py
@@ -1,15 +1,0 @@
-""" __init__.py
-Description:
-"""
-# Package Header #
-from ..header import *
-
-# Header #
-__author__ = __author__
-__credits__ = __credits__
-__maintainer__ = __maintainer__
-__email__ = __email__
-
-# Imports #
-# Local Packages #
-from .hdf5eegframe import HDF5BaseFrame
